+++ quill-core/src/main/scala/io/getquill/ast/StatelessTransformer.scala
@@ -11,7 +11,7 @@ trait StatelessTransformer {
       case e: Assignment           => apply(e)
       case Function(params, body)  => Function(params, apply(body))
       case e: Ident                => e
-      case Property(a, name)       => Property(apply(a), name)
+      case e: Property             => apply(e)
       case Infix(a, b)             => Infix(a, b.map(apply))
       case e: OptionOperation      => apply(e)
       case e: TraversableOperation => apply(e)
@@ -22,17 +22,22 @@ trait StatelessTransformer {
       case Block(statements)       => Block(statements.map(apply))
       case Val(name, body)         => Val(name, apply(body))
       case o: Ordering             => o
+      case e: OnConflict.Excluded  => e
+      case e: OnConflict.Existing  => e
     }
 
   def apply(o: OptionOperation): OptionOperation =
     o match {
-      case OptionMap(a, b, c)    => OptionMap(apply(a), b, apply(c))
-      case OptionForall(a, b, c) => OptionForall(apply(a), b, apply(c))
-      case OptionExists(a, b, c) => OptionExists(apply(a), b, apply(c))
-      case OptionContains(a, b)  => OptionContains(apply(a), apply(b))
-      case OptionIsEmpty(a)      => OptionIsEmpty(apply(a))
-      case OptionNonEmpty(a)     => OptionNonEmpty(apply(a))
-      case OptionIsDefined(a)    => OptionIsDefined(apply(a))
+      case OptionFlatten(a)       => OptionFlatten(apply(a))
+      case OptionGetOrElse(a, b)  => OptionGetOrElse(apply(a), apply(b))
+      case OptionFlatMap(a, b, c) => OptionFlatMap(apply(a), b, apply(c))
+      case OptionMap(a, b, c)     => OptionMap(apply(a), b, apply(c))
+      case OptionForall(a, b, c)  => OptionForall(apply(a), b, apply(c))
+      case OptionExists(a, b, c)  => OptionExists(apply(a), b, apply(c))
+      case OptionContains(a, b)   => OptionContains(apply(a), apply(b))
+      case OptionIsEmpty(a)       => OptionIsEmpty(apply(a))
+      case OptionNonEmpty(a)      => OptionNonEmpty(apply(a))
+      case OptionIsDefined(a)     => OptionIsDefined(apply(a))
     }
 
   def apply(o: TraversableOperation): TraversableOperation =
@@ -69,6 +74,11 @@ trait StatelessTransformer {
       case Assignment(a, b, c) => Assignment(a, apply(b), apply(c))
     }
 
+  def apply(e: Property): Property =
+    e match {
+      case Property(a, name) => Property(apply(a), name)
+    }
+
   def apply(e: Operation): Operation =
     e match {
       case UnaryOperation(o, a)            => UnaryOperation(o, apply(a))
@@ -94,6 +104,19 @@ trait StatelessTransformer {
       case Delete(query)                     => Delete(apply(query))
       case Returning(query, alias, property) => Returning(apply(query), alias, apply(property))
       case Foreach(query, alias, body)       => Foreach(apply(query), alias, apply(body))
+      case OnConflict(query, target, action) => OnConflict(apply(query), apply(target), apply(action))
+    }
+
+  def apply(e: OnConflict.Target): OnConflict.Target =
+    e match {
+      case OnConflict.NoTarget          => e
+      case OnConflict.Properties(props) => OnConflict.Properties(props.map(apply))
+    }
+
+  def apply(e: OnConflict.Action): OnConflict.Action =
+    e match {
+      case OnConflict.Ignore          => e
+      case OnConflict.Update(assigns) => OnConflict.Update(assigns.map(apply))
     }
 
 }