+++ quill-sql/src/test/sql/postgres-schema.sql
@@ -60,7 +60,7 @@ CREATE TABLE EncodingUUIDTestEntity(
 
 CREATE TABLE TestEntity(
     s VARCHAR(255),
-    i INTEGER,
+    i INTEGER primary key,
     l BIGINT,
     o INTEGER
 );